Add storage provider
